uc_set_native_thunks: generate TBs for PCs identified as native code
This provides a more efficient mechanism for thunking to native code than:
- exits
- UC_HOOK_FIND_TB_FAILURE
- permission faults

Only supports i386 target today.

This is the only way to get any speed without the notdirty_write fix.
With the notdirty_write fix, the benefit is less pronounced but is
still real.

Not used by MultiArchUefiPkg because UC is not really reentrance-safe.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>