uc_invalidate_tb: fix correctness
This is not aware of the underlying RAM block organization.
It's not valid to invalidate the linear range of addresses
starting at get_page_addr_code, as the mapping is not linear,
esp after the kind of fragmentation that happens after
calls to uc_mem_protect.

The only safe thing to do here is to go page by page.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>